+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 
 - Non-superusers in the admin module can now see system rules.
 
+- The Alias-model no longer uses recursion to clean up connections to 
+  DocumentReports.
+
 ## Version 3.22.2, 12th February 2024
 
 "Unparalleled Spectacles"